requirements.txt
aiogram==3.3.0
cryptography==41.0.7
bcrypt==4.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1